pyproject.toml
[tool.poetry]
name = "pyspark_pg"
version = "0.1.0"
description = ""
authors = ["Anant Aneja <anant@ahana.io>"]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = {extras = ["sql"], version = "^3.2.0"}
delta-spark = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
